requirements.txt
tensorflow
opencv-python
numpy
pillow
customtkinter
imageio-ffmpeg
requests

